+++ NEWS.md
@@ -1,5 +1,7 @@
 # Changes in version 1.0.4 (release date:X-X-2025)
 
+- Fitbit: Fix bug preventing the loading of a sequence of json files. #76
+
 - Add functions for reading Parmay Tech Matrix sensor (bin/BIN files) with accelerometer, gyroscope, temperature, and heart rate #70.
 
 # Changes in version 1.0.3 (release date:07-03-2025)